requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
python-dateutil==2.8.2
APScheduler==3.10.4
requests==2.32.3
PyJWT==2.9.0
